🚚 Rename test sample data to fixtures
