+++ src/components/__tests__/__fixtures__/sampleFlawAffects.ts
