+++ src/components/__tests__/__fixtures__/sampleFlawCvssScores.ts
